Move thread rest inside thread
