+++ .agents/skills/ak-dev-architecture/SKILL.md
@@ -202,7 +202,7 @@ Provides persistent, named conversation threads keyed by `session_id`, gated beh
 - **`Thread` / `ThreadMessage` / `ThreadAttachment` / `ThreadPage` / `MessagePage`** (`model.py`): Pydantic models for thread metadata, individual messages, attachment references, and cursor-paginated listings
 - **`ThreadNamingStrategy`** (`naming.py`): Overridable strategy that names auto-created threads — default implementation makes a single LiteLLM call (`thread.naming.model`, requires the `thread` extra) to derive a concise title from the first prompt, falling back to a truncated prompt prefix when `litellm`/an API key is unavailable. Explicit `thread_name` on a chat request always wins and locks the thread against further automatic naming
 - **`Authoriser`** (`authoriser.py`): Pluggable base class (`authorise(token) -> Optional[user_id]`) that `ThreadRESTRequestHandler` calls to protect the read routes; routes are open when no `Authoriser` is configured
-- **`ThreadRESTRequestHandler`** (`api/thread.py`): Mounts `GET /api/v1/threads` (list, filterable by `user_id`/`group_id`, cursor-paginated) and `GET /api/v1/threads/{session_id}` (thread + paginated message history); raises 404 when thread support is disabled and 403 when a resolved `user_id` doesn't own the requested thread. **Never mounted implicitly** — it is served only when passed to `RESTAPI.run(handlers=[...])`, so a `thread:` block configures storage alone and enabling a backend cannot silently publish a read API over every user's history. Consequence: `ECSIOHandler.run()` and `CloudRun.run()` build their handler lists internally and therefore cannot expose thread routes today (`RESTAPI.add()` is not a substitute — it mounts under `api.custom_router_prefix`, giving the wrong paths)
+- **`ThreadRESTRequestHandler`** (`rest.py`): Mounts `GET /api/v1/threads` (list, filterable by `user_id`/`group_id`, cursor-paginated) and `GET /api/v1/threads/{session_id}` (thread + paginated message history); raises 404 when thread support is disabled and 403 when a resolved `user_id` doesn't own the requested thread. **Never mounted implicitly** — it is served only when passed to `RESTAPI.run(handlers=[...])`, so a `thread:` block configures storage alone and enabling a backend cannot silently publish a read API over every user's history. Consequence: `ECSIOHandler.run()` and `CloudRun.run()` build their handler lists internally and therefore cannot expose thread routes today (`RESTAPI.add()` is not a substitute — it mounts under `api.custom_router_prefix`, giving the wrong paths)
 
 ### Store Backends
