+++ ak-py/src/agentkernel/api/__init__.py
@@ -13,4 +13,3 @@ except importlib.metadata.PackageNotFoundError:
 
 from .handler import AgentRESTRequestHandler, RESTRequestHandler
 from .http import RESTAPI
-from .thread import ThreadRESTRequestHandler
